backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pillow==10.4.0
python-dotenv==1.0.1
pydantic[email]==2.9.2
ultralytics==8.3.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
